requirements.txt
pdfplumber==0.8.0
pandas==1.5.3
matplotlib==3.7.1
streamlit==1.20.0
scikit-learn==1.2.0
seaborn==0.12.2
torch==2.1.0
transformers==4.28.1
numpy==1.24.2
pytesseract==0.3.9
opencv-python-headless==4.6.0.66
Pillow==9.4.0
scipy==1.10.0
python-dotenv==1.0.0
